+++ myems-admin/app/controllers/settings/hybridpowerstation/hybridpowerstation.master.controller.js
@@ -1,0 +1,10 @@
+'use strict';
+
+app.controller('HybridPowerStationMasterController', function($scope) {
+
+
+	$scope.$on('handleEmitHybridPowerStationChanged', function(event) {
+		$scope.$broadcast('handleBroadcastHybridPowerStationChanged');
+	});
+
+});